+++ common/gen_lag_kern.m
@@ -29,7 +29,7 @@
 % John M. O' Toole, University College Cork
 % Started: 16-04-2014
 %
-% last update: Time-stamp: <2014-07-23 13:04:21 (otoolej)>
+% last update: Time-stamp: <2021-08-24 17:42:46 (otoolej)>
 %-------------------------------------------------------------------------------
 function [g2,P,Ph_floor,Nfreq]=gen_lag_kern(win_params,N,Nfreq)
 if(nargin<3 || isempty(Nfreq)), Nfreq=[]; end
@@ -77,7 +77,7 @@ end
 % 2. call function to generate window
 %---------------------------------------------------------------------
 win_type=win_params{2}; 
-if( l>=3 ) win_extra_param=win_params{3}; else, win_extra_param=0; end
+if( l>=3 ) win_extra_param=win_params{3}; else, win_extra_param=[]; end
 if( l>3 )  win_dft_param=win_params{4}; else, win_dft_param=0; end
 
 
@@ -110,6 +110,12 @@ if(DBplot)
     subplot(212); plot(g2);    
 end
 
+
+function w=shiftWin(w)
+%---------------------------------------------------------------------
+% shift window to centre
+%---------------------------------------------------------------------
+w=circshift(w(:),ceil(length(w)/2));